animation/quote
Added animation to portfolio section and added a quote to the image break

My plan for services section is to have each respective button change the text to relative information about the service. I've considered the idea of doing this in react but I think it may be a bit overkill.